Move setup_logging() to utils
