+++ p2a_impacts/resolver.py
@@ -6,18 +6,7 @@ import logging
 from .parser import build_parse_tree
 from .evaluator import evaluate_rule
 from .fetch_data import get_dict_val, read_csv, get_variables
-
-
-def setup_logging(log_level):
-    formatter = logging.Formatter(
-        "%(asctime)s %(levelname)s: %(message)s", "%Y-%m-%d %H:%M:%S"
-    )
-    handler = logging.StreamHandler()
-    handler.setFormatter(formatter)
-    logger = logging.getLogger("scripts")
-    logger.addHandler(handler)
-    logger.setLevel(getattr(logging, log_level))
-    return logger
+from .utils import setup_logging
 
 
 def resolve_rules(
